move DebugStack into collect module
